+++ integer-reversal/main.go
@@ -3,6 +3,6 @@ package main
 func main() {
 }
 
-func reverse(n int16) (result int16) {
+func reverse(n int) (result int) {
 
 }
